An 8 oscillator synth i think? (7 main osc, 1 sub, 1 noise)
It's pretty simple, every osc has it's own adsr and Volume and Pan, and are divided into 2 groups, which also have Volume and Pan controls:
- Basic Osc (Pulse, Saw, Sine, Triangle)
- Special Osc (Wavetable, Phase1, Swarm)

Also features a filter (Controllable in Performance Tab) Post Master Mixer and Pre Master ADSR.

Everything is controllable in the Modulator tab of the synth. (Mainly Macro-4 and ADSR) to make it the most front end possible.