https://i.imgur.com/LRQOkGp.png
This was originally an attempt of re-creating the "Fresh Air" VST with just stock devices in Bitwig.
But I ended up giving it tons of extra features.

So I call this "Ultimate vocal tool".

Put it on your vocals, and they will sound crisp, clear and "up front".

Here's what it sounds like:

Dry vocals:
https://soundcloud.com/sebastianrobert/vocal-dry/s-ZSHx3F8DaS6

Wet vocals using this device:
https://soundcloud.com/sebastianrobert/vocal-wet/s-KykSfw1S61r