*(V2 - Fixes an inconsistent mapping with the "Color" knob)*
The real and true Erosion sound from Ableton, in Bitwig at last! This is a custom rack I built using a free VST that somebody finally made and got the Erosion process and sound nearly identical to the original. What they left out was the "Wide" noise mode, so I created this rack to enable the Wide mode too.  You'll need the VST itself for this rack to work. The VST is free and you can find the DL link in the rack preset's description. See the production  channel for a video demo.