<@360540101853446165>  hera ya go.
for the uninitiated, it's a kick generation tool made in the grid. it uses 4 sample layers, a noise layer, and a noise transient going through filters with most parameters linked to dices which are linked to sub group triggers, which are linked to a master trigger, meaning you can either press one button to randomize everything, or press one trigger and randomize only the pitch envelopes, filter positions, sample pitches, etc., etc. by only clicking the corresponding trigger group.