This is a linear-phase five band approximation of the Loud Split device using robbert-vdh's crossover plugin (https://github.com/robbert-vdh/nih-plug/tree/master/plugins/crossover). Lo Thresh corresponds to the divide between the quiet and mid volume frequencies, and Hi Thresh corresponds to the divide between the mid and high volume frequencies. Put your effects in the "FX HERE" FX layer at the end of the corresponding chain (high, mid, or low).
Advanced: There are controls for the number of bands and the location of each crossover point, plus you can disable the linear phase processing if you wish. Additionally, you can change the depth, attack, and release controls for the army of gate devices powering this behind the scenes, though generally you want to leave depth all the way up and the attack/release macros very low.